requirements.txt
streamlit==1.35.0
pillow==10.3.0
openai==1.24.0
langchain==0.2.0
python-dotenv==1.0.1
